+++ examples/base_get_request/main.go
@@ -1,3 +1,4 @@
+// Package main shows a basic GET request with MaiGo.
 package main
 
 import (